[#852532] Exception when editing node_kaltura_entry types and remote media is not found.
Move kaltura_get_entry() to a separate function so we don't duplicate code in various submodules.

Might want to add a static cache to this function so we can reduce API hits but this is a good start IMO.

Validate that entry exists in node_kaltura_entry_validate() and field_kaltura_field('validate').